Increase waypoint following robustness
In the waypoint_follower_core.cpp's verifyFollowing() function,
the code checks whether the car has deviated "enough" from the target
waypoint. (Only) if it doesn't, new adjusted twist commands are
issued. This commit lowers the thresholds so that the car should wander
less in the lane and stick to the actual waypoints a bit more.

This issue has been discussed in the DBW walkhtrough video, with one
suggested solution being to always force the node to follow the
waypoints, no matter what. I didn't want to go to this extreme though.