Add constraint for number of feet
Before this commit, a (virtual) driver could simultaneously
operate brake and throttle, effectively telling the car to speed up
and slow down at the same time.

While there might be actual reasons to do this, this commit sets
the throttle value to zero whenever a nonzero brake command is issued.